+++ CHANGELOG.md
@@ -10,6 +10,7 @@ All notable changes to this project will be documented in this file.
 ### Breaking Changed
 
 ### Changed
+- Berry make Leds animate calls reentrant
 
 ### Fixed
 - ESP32 rules operation priority regression from v13.3.0.4 (#22636)